Show default index url provided by pip
